refactored and moved things around
