+++ old/json_to_predicate_calculus.py
